requirements.txt
requests
pandas
matplotlib
seaborn
sqlalchemy
psycopg2-binary
python-dotenv